T-Shirt Vendor
Randy is a t-shirt vendor who plans to order t-shirts with the names of the NCAA final four teams from a manufacturer and then sell them to fans.  The fixed cost of any order is $750, the variable cost per T-shirt to Randy is $8, and Randy’s selling price is $18.  However, this price will be charged only until a week after the tournament. After that time, Randy figures that the interest in the t-shirts will be low, so he plans to sell all remaining t-shirts, if any, at $6 each.  His best guess is that the demand for the t-shirts during the full price period will be 1500.  He is thinking about ordering 1450 t-shirts, but he wants to build a spreadsheet model that will let him experiment with the uncertain demand and his order quantity.