Woodworks Bookshelf Co.
The Woodworks Company produces a variety of custom-designed wood furniture, made from either cherry or oak. The company knows that wood prices and labor costs are likely to increase in the future. The table below shows the number of board-feet and labor hours required for a bookshelf, the current costs per board-foot and labor hour, and the anticipated annual increases in these costs.
Build a spreadsheet model that enables the company to experiment with the growth rates in wood and labor costs so that a manager can see, both numerically and graphically, how the costs of bookshelves vary in the next few years.